Implementacion_Nube/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.3.0
fastapi==0.116.2
grpcio==1.75.0
grpcio-tools==1.75.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
python-multipart==0.0.20
requests==2.7.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
wheel==0.45.1